hrz-2980: init refactoring
