+++ libs/domain/search/facet/facet.component.spec.ts
