+++ libs/domain/search/facet/index.ts
@@ -1,1 +1,4 @@
-export * from './src';
+export * from './facet.controller';
+export * from './facet.component';
+export * from './facet.model';
+export * from './facet.styles';
