--- libs/domain/search/facet/src/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './facet.controller';
