--- libs/domain/search/facet/src/index.ts
@@ -1,4 +1,0 @@
-export * from './controllers';
-export * from './facet.component';
-export * from './facet.model';
-export * from './facet.styles';
